requirements.txt
audioread>=2.1.9
certifi>=2021.5.30
cffi>=1.15.0
charset-normalizer>=2.0.0
decorator>=4.4.2
idna>=2.10
joblib>=1.0.1
lazy_loader>=0.1
librosa>=0.8.1
llvmlite>=0.36.0
msgpack>=1.0.2
numba>=0.53.1
numpy>=1.20.3
openai>=0.27.0
packaging>=20.9
pandas>=1.3.0
platformdirs>=2.5.0
pooch>=1.4.0
pycparser>=2.20
python-dateutil>=2.8.1
python-dotenv>=0.19.0
pytz>=2021.1
PyYAML>=6.0
requests>=2.25.1
scikit-learn>=0.24.2
scipy>=1.7.0
six>=1.16.0
sounddevice>=0.4.1
soundfile>=0.10.3
soxr>=0.3.2
threadpoolctl>=2.1.0
tqdm>=4.61.1
typing_extensions>=3.10.0
urllib3>=1.26.5
pydub>=0.25.1